requirements.txt
tensorflow
pillow
numpy
streamlit
